Move PlaylistConfig to the correct dir
